validate change password and add bottom sheet logout
