--- lib/views/screens/auth/Login/newLogin.dart
@@ -1,262 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:health_care/common/app_colors.dart';
-import 'package:health_care/viewmodels/auth_viewmodel.dart';
-import 'package:health_care/views/screens/auth/Login/forgotPassword_screen.dart';
-import 'package:health_care/views/screens/auth/Login/newRes.dart';
-import 'package:provider/provider.dart';
-
-class LoginScreen extends StatefulWidget {
-  const LoginScreen({Key? key}) : super(key: key);
-
-  @override
-  State<LoginScreen> createState() => _LoginScreenState();
-}
-
-class _LoginScreenState extends State<LoginScreen> {
-  final _formKey = GlobalKey<FormState>();
-  final TextEditingController _emailOrPhoneController = TextEditingController();
-  final TextEditingController _passwordController = TextEditingController();
-  bool _obscurePassword = true;
-  bool _isLoading = false;
-
-  /// Hàm kiểm tra Gmail hoặc SĐT
-  String? _validateEmailOrPhone(String? value) {
-    if (value == null || value.isEmpty) {
-      return 'Vui lòng nhập Gmail hoặc SĐT';
-    }
-
-    final emailRegex = RegExp(r'^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$');
-    final phoneRegex = RegExp(r'^(0|\+84)[0-9]{9}$');
-
-    if (!emailRegex.hasMatch(value) && !phoneRegex.hasMatch(value)) {
-      return 'Sai định dạng Gmail hoặc SĐT';
-    }
-    return null;
-  }
-
-  /// Hàm kiểm tra mật khẩu
-  String? _validatePassword(String? value) {
-    if (value == null || value.isEmpty) {
-      return 'Vui lòng nhập mật khẩu';
-    }
-    return null;
-  }
-
-  void _togglePasswordVisibility() {
-    setState(() {
-      _obscurePassword = !_obscurePassword;
-    });
-  }
-
-  /// Hàm đăng nhập sử dụng AuthViewModel
-  void _signIn() async {
-    if (_formKey.currentState!.validate()) {
-      setState(() {
-        _isLoading = true;
-      });
-
-      final authViewModel = Provider.of<AuthViewModel>(context, listen: false);
-      String phoneOrEmail = _emailOrPhoneController.text.trim();
-      String password = _passwordController.text.trim();
-
-      try {
-        await authViewModel.login(context, phoneOrEmail, password);
-      } finally {
-        setState(() {
-          _isLoading = false;
-        });
-      }
-    }
-  }
-
-  void _signUp() {
-    Navigator.push(
-      context,
-      MaterialPageRoute(
-          builder: (context) =>
-              RegisterScreen()), // Thay thế bằng màn hình đăng ký của bạn
-    );
-  }
-
-  void _forgotPassword() {
-    Navigator.push(
-      context,
-      MaterialPageRoute(builder: (context) => ForgotPasswordScreen()),
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Stack(
-        children: [
-          Container(
-            height: double.infinity,
-            width: double.infinity,
-            decoration: const BoxDecoration(
-              gradient: LinearGradient(
-                colors: [
-                  AppColors.deepBlue,
-                  Color.fromARGB(255, 210, 231, 243)
-                ], // Màu bắt đầu và kết thúc
-              ),
-            ),
-            child: const Padding(
-                padding: EdgeInsets.only(top: 60.0, left: 22),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      'Xin mời',
-                      style: TextStyle(
-                          fontSize: 20,
-                          color: Colors.white,
-                          fontWeight: FontWeight.bold),
-                    ),
-                    Text(
-                      'Đăng nhập',
-                      style: TextStyle(
-                          fontSize: 30,
-                          color: Colors.white,
-                          fontWeight: FontWeight.bold),
-                    ),
-                  ],
-                )),
-          ),
-          Padding(
-            padding: const EdgeInsets.only(top: 250.0),
-            child: Container(
-              decoration: const BoxDecoration(
-                borderRadius: BorderRadius.only(
-                    topLeft: Radius.circular(50),
-                    topRight: Radius.circular(50)),
-                color: Colors.white,
-              ),
-              height: double.infinity,
-              width: double.infinity,
-              child: Padding(
-                padding:
-                    const EdgeInsets.symmetric(horizontal: 18, vertical: 50),
-                child: SingleChildScrollView(
-                  child: Form(
-                    key: _formKey,
-                    child: Column(
-                      children: [
-                        TextFormField(
-                          controller: _emailOrPhoneController,
-                          keyboardType: TextInputType.emailAddress,
-                          validator: _validateEmailOrPhone,
-                          decoration: const InputDecoration(
-                            suffixIcon: Icon(Icons.check, color: Colors.grey),
-                            label: Text(
-                              'Nhập email hoặc số điện thoại',
-                              style: TextStyle(
-                                  fontWeight: FontWeight.bold,
-                                  color: Colors.grey,
-                                  fontSize: 14),
-                            ),
-                          ),
-                        ),
-                        const SizedBox(height: 10),
-                        TextFormField(
-                          controller: _passwordController,
-                          obscureText: _obscurePassword,
-                          validator: _validatePassword,
-                          decoration: InputDecoration(
-                            suffixIcon: IconButton(
-                              icon: Icon(
-                                _obscurePassword
-                                    ? Icons.visibility_off
-                                    : Icons.visibility,
-                                color: Colors.grey,
-                              ),
-                              onPressed: _togglePasswordVisibility,
-                            ),
-                            label: const Text(
-                              'Nhập mật khẩu',
-                              style: TextStyle(
-                                  fontWeight: FontWeight.bold,
-                                  color: Colors.grey,
-                                  fontSize: 14),
-                            ),
-                          ),
-                        ),
-                        const SizedBox(height: 10),
-                        Align(
-                          alignment: Alignment.bottomLeft,
-                          child: GestureDetector(
-                            onTap: _forgotPassword,
-                            child: Text(
-                              'Quên mật khẩu',
-                              style: TextStyle(
-                                fontWeight: FontWeight.bold,
-                                fontSize: 13,
-                                color: Color(0xff281537),
-                              ),
-                            ),
-                          ),
-                        ),
-                        const SizedBox(height: 100),
-                        GestureDetector(
-                          onTap: _isLoading ? null : _signIn,
-                          child: Container(
-                            height: 55,
-                            width: 300,
-                            decoration: BoxDecoration(
-                              borderRadius: BorderRadius.circular(30),
-                              gradient: const LinearGradient(colors: [
-                                AppColors.deepBlue,
-                                AppColors.softBlue,
-                              ]),
-                            ),
-                            child: Center(
-                              child: _isLoading
-                                  ? CircularProgressIndicator(
-                                      color: Colors.white,
-                                    )
-                                  : Text(
-                                      'Đăng nhập',
-                                      style: TextStyle(
-                                          fontWeight: FontWeight.bold,
-                                          fontSize: 20,
-                                          color: Colors.white),
-                                    ),
-                            ),
-                          ),
-                        ),
-                        const SizedBox(height: 10),
-                        Row(
-                          crossAxisAlignment: CrossAxisAlignment.center,
-                          mainAxisAlignment: MainAxisAlignment.center,
-                          children: [
-                            const Text(
-                              "Bạn chưa có tài khoản?",
-                              style: TextStyle(
-                                  fontWeight: FontWeight.w500,
-                                  color: Colors.grey),
-                            ),
-                            SizedBox(width: 5),
-                            GestureDetector(
-                              onTap: _signUp,
-                              child: Text(
-                                "Đăng ký",
-                                style: TextStyle(
-                                    fontWeight: FontWeight.bold,
-                                    fontSize: 15,
-                                    color: AppColors.deepBlue),
-                              ),
-                            ),
-                          ],
-                        ),
-                      ],
-                    ),
-                  ),
-                ),
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
