+++ lib/views/screens/profile/change_password.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:fluttertoast/fluttertoast.dart';
 import 'package:health_care/common/app_colors.dart';
 import 'package:health_care/viewmodels/auth_viewmodel.dart';
 import 'package:health_care/views/widgets/widget_header_body.dart';
@@ -17,18 +16,12 @@ class _ChangePasswordScreenState extends State<ChangePasswordScreen> {
   final TextEditingController _newPasswordController = TextEditingController();
   final TextEditingController _confirmPasswordController =
       TextEditingController();
+  final _formKey = GlobalKey<FormState>(); // To validate form
+
   bool _isOldPasswordVisible = false;
   bool _isNewPasswordVisible = false;
   bool _isConfirmPasswordVisible = false;
-  bool isButtonEnabled = false;
-
-  @override
-  void initState() {
-    super.initState();
-    _oldPasswordController.addListener(_validateForm);
-    _newPasswordController.addListener(_validateForm);
-    _confirmPasswordController.addListener(_validateForm);
-  }
+  bool isLoading = false;
 
   @override
   void dispose() {
@@ -38,61 +31,51 @@ class _ChangePasswordScreenState extends State<ChangePasswordScreen> {
     super.dispose();
   }
 
-  void _validateForm() {
-    setState(() {
-      isButtonEnabled = _oldPasswordController.text.isNotEmpty &&
-          _newPasswordController.text.isNotEmpty &&
-          _confirmPasswordController.text.isNotEmpty;
-    });
+  String? _validateOldPassword(String? value) {
+    if (value == null || value.trim().isEmpty) {
+      return 'Mật khẩu cũ không được trống';
+    }
+    return null;
+  }
+
+  String? _validateNewPassword(String? value) {
+    if (value == null || value.trim().isEmpty) {
+      return 'Mật khẩu mới không được trống';
+    }
+    if (value.length < 6) {
+      return 'Mật khẩu mới phải có ít nhất 6 ký tự';
+    }
+    return null;
+  }
+
+  String? _validateConfirmPassword(String? value) {
+    if (value == null || value.trim().isEmpty) {
+      return 'Xác nhận mật khẩu không được trống';
+    }
+    if (value != _newPasswordController.text) {
+      return 'Mật khẩu xác nhận không khớp';
+    }
+    return null;
   }
 
   void _handleChangePassword() async {
-    final authViewModel = Provider.of<AuthViewModel>(context, listen: false);
+    if (_formKey.currentState!.validate()) {
+      setState(() {
+        isLoading = true;
+      });
 
-    String oldPassword = _oldPasswordController.text.trim();
-    String newPassword = _newPasswordController.text.trim();
-    String confirmPassword = _confirmPasswordController.text.trim();
+      String oldPassword = _oldPasswordController.text.trim();
+      String newPassword = _newPasswordController.text.trim();
+      String confirmPassword = _confirmPasswordController.text.trim();
 
+      final authViewModel = Provider.of<AuthViewModel>(context, listen: false);
+      await authViewModel.changePassword(
+          context, oldPassword, newPassword, confirmPassword);
 
-    //   // Gọi API kiểm tra mật khẩu cũ có đúng không
-    // bool isOldPasswordCorrect = await authViewModel.verifyOldPassword(oldPassword);
-
-    // if (!isOldPasswordCorrect) {
-    //   Fluttertoast.showToast(
-    //     msg: "Mật khẩu cũ không chính xác.",
-    //     toastLength: Toast.LENGTH_LONG,
-    //     gravity: ToastGravity.BOTTOM,
-    //     backgroundColor: Colors.red,
-    //     textColor: Colors.white,
-    //   );
-    //   return;
-    // }
-
-
-    if (newPassword == oldPassword) {
-      Fluttertoast.showToast(
-        msg: "Mật khẩu mới không được trùng với mật khẩu cũ.",
-        toastLength: Toast.LENGTH_SHORT,
-        gravity: ToastGravity.BOTTOM,
-        backgroundColor: Colors.red,
-        textColor: Colors.white,
-      );
-      return;
+      setState(() {
+        isLoading = false;
+      });
     }
-
-    if (newPassword != confirmPassword) {
-      Fluttertoast.showToast(
-        msg: "Mật khẩu mới và xác nhận mật khẩu không khớp.",
-        toastLength: Toast.LENGTH_SHORT,
-        gravity: ToastGravity.BOTTOM,
-        backgroundColor: Colors.red,
-        textColor: Colors.white,
-      );
-      return;
-    }
-
-    await authViewModel.changePassword(
-        context, oldPassword, newPassword, confirmPassword);
   }
 
   @override
@@ -100,88 +83,98 @@ class _ChangePasswordScreenState extends State<ChangePasswordScreen> {
     return WidgetHeaderBody(
       iconBack: true,
       title: 'Đổi Mật Khẩu',
-      body: Padding(
-        padding: const EdgeInsets.all(16.0),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            _customTitle(title: 'Mật khẩu cũ'),
-            _buildPasswordField(
-              controller: _oldPasswordController,
-              label: 'Nhập mật khẩu cũ',
-              isVisible: _isOldPasswordVisible,
-              onToggle: () {
-                setState(() {
-                  _isOldPasswordVisible = !_isOldPasswordVisible;
-                });
-              },
-            ),
-            const SizedBox(height: 16),
-            _customTitle(title: 'Mật khẩu mới'),
-            _buildPasswordField(
-              controller: _newPasswordController,
-              label: 'Nhập mật khẩu mới',
-              isVisible: _isNewPasswordVisible,
-              onToggle: () {
-                setState(() {
-                  _isNewPasswordVisible = !_isNewPasswordVisible;
-                });
-              },
-            ),
-            const SizedBox(height: 16),
-            _customTitle(title: 'Nhập lại mật khẩu mới'),
-            _buildPasswordField(
-              controller: _confirmPasswordController,
-              label: 'Nhập lại mật khẩu mới',
-              isVisible: _isConfirmPasswordVisible,
-              onToggle: () {
-                setState(() {
-                  _isConfirmPasswordVisible = !_isConfirmPasswordVisible;
-                });
-              },
-            ),
-            const SizedBox(height: 24),
-            SizedBox(
-              width: double.infinity,
-              child: OutlinedButton(
-                onPressed: isButtonEnabled ? _handleChangePassword : null,
-                style: OutlinedButton.styleFrom(
-                  side: BorderSide(
-                      color: isButtonEnabled
-                          ? AppColors.deepBlue
-                          : AppColors.grey4),
-                  backgroundColor:
-                      isButtonEnabled ? AppColors.deepBlue : AppColors.grey4,
-                  shape: const RoundedRectangleBorder(
-                    borderRadius: BorderRadius.all(Radius.circular(10)),
+      body: SingleChildScrollView(
+        child: Padding(
+          padding: const EdgeInsets.all(16.0),
+          child: Form(
+            key: _formKey,
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                _customTitle(title: 'Mật khẩu cũ'),
+                _buildPasswordField(
+                  controller: _oldPasswordController,
+                  label: 'Nhập mật khẩu cũ',
+                  isVisible: _isOldPasswordVisible,
+                  onToggle: () {
+                    setState(() {
+                      _isOldPasswordVisible = !_isOldPasswordVisible;
+                    });
+                  },
+                  validator: _validateOldPassword,
+                ),
+                const SizedBox(height: 15),
+                _customTitle(title: 'Mật khẩu mới'),
+                _buildPasswordField(
+                  controller: _newPasswordController,
+                  label: 'Nhập mật khẩu mới',
+                  isVisible: _isNewPasswordVisible,
+                  onToggle: () {
+                    setState(() {
+                      _isNewPasswordVisible = !_isNewPasswordVisible;
+                    });
+                  },
+                  validator: _validateNewPassword,
+                ),
+                const SizedBox(height: 15),
+                _customTitle(title: 'Xác nhận mật khẩu'),
+                _buildPasswordField(
+                  controller: _confirmPasswordController,
+                  label: 'Nhập lại mật khẩu',
+                  isVisible: _isConfirmPasswordVisible,
+                  onToggle: () {
+                    setState(() {
+                      _isConfirmPasswordVisible = !_isConfirmPasswordVisible;
+                    });
+                  },
+                  validator: _validateConfirmPassword,
+                ),
+                const SizedBox(height: 15),
+                SizedBox(
+                  width: double.infinity,
+                  child: OutlinedButton(
+                    onPressed: isLoading ? null : _handleChangePassword,
+                    style: OutlinedButton.styleFrom(
+                      side: BorderSide(
+                          color: !isLoading
+                              ? AppColors.deepBlue
+                              : AppColors.grey4),
+                      backgroundColor:
+                          !isLoading ? AppColors.deepBlue : AppColors.grey4,
+                      shape: const RoundedRectangleBorder(
+                        borderRadius: BorderRadius.all(Radius.circular(25)),
+                      ),
+                      padding: const EdgeInsets.symmetric(vertical: 12),
+                    ),
+                    child: isLoading
+                        ? CircularProgressIndicator(
+                            valueColor: AlwaysStoppedAnimation<Color>(
+                                AppColors.neutralWhite),
+                          )
+                        : Text(
+                            'Xác nhận',
+                            style: TextStyle(
+                                fontSize: 16,
+                                fontWeight: FontWeight.bold,
+                                color: AppColors.neutralWhite),
+                          ),
                   ),
-                  padding: const EdgeInsets.symmetric(vertical: 15),
                 ),
-                child: Text(
-                  'Xác nhận',
-                  style: TextStyle(
-                      fontSize: 17,
-                      fontWeight: FontWeight.bold,
-                      color: AppColors.neutralWhite),
-                ),
-              ),
+              ],
             ),
-          ],
+          ),
         ),
       ),
     );
   }
 
   Widget _customTitle({required String title}) {
-    return Container(
-      margin: const EdgeInsets.only(bottom: 8),
-      child: Text(
-        title,
-        style: TextStyle(
-          fontSize: 14,
-          fontWeight: FontWeight.bold,
-          color: AppColors.deepBlue,
-        ),
+    return Text(
+      title,
+      style: TextStyle(
+        fontSize: 14,
+        fontWeight: FontWeight.bold,
+        color: Colors.black,
       ),
     );
   }
@@ -191,31 +184,35 @@ class _ChangePasswordScreenState extends State<ChangePasswordScreen> {
     required String label,
     required bool isVisible,
     required VoidCallback onToggle,
+    required String? Function(String?) validator,
   }) {
-    return Container(
-      height: 45,
-      margin: const EdgeInsets.symmetric(vertical: 5),
-      child: TextField(
-        controller: controller,
-        obscureText: !isVisible,
-        decoration: InputDecoration(
-          labelText: label,
-          filled: true,
-          fillColor: Colors.white,
-          floatingLabelBehavior: FloatingLabelBehavior.never,
-          labelStyle: const TextStyle(fontSize: 14),
-          contentPadding:
-              const EdgeInsets.symmetric(vertical: 10, horizontal: 15),
-          border: OutlineInputBorder(
-            borderRadius: const BorderRadius.all(Radius.circular(10)),
-            borderSide: BorderSide(color: AppColors.accent, width: 1),
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        TextFormField(
+          controller: controller,
+          obscureText: !isVisible,
+          decoration: InputDecoration(
+            labelText: label,
+            filled: true,
+            fillColor: Colors.white,
+            floatingLabelBehavior: FloatingLabelBehavior.never,
+            labelStyle:
+                const TextStyle(fontSize: 14, color: AppColors.neutralDark),
+            contentPadding:
+                const EdgeInsets.symmetric(vertical: 10, horizontal: 15),
+            border: OutlineInputBorder(
+              borderRadius: const BorderRadius.all(Radius.circular(10)),
+              borderSide: BorderSide(color: AppColors.accent, width: 1),
+            ),
+            suffixIcon: IconButton(
+              icon: Icon(isVisible ? Icons.visibility : Icons.visibility_off),
+              onPressed: onToggle,
+            ),
           ),
-          suffixIcon: IconButton(
-            icon: Icon(isVisible ? Icons.visibility : Icons.visibility_off),
-            onPressed: onToggle,
-          ),
+          validator: validator,
         ),
-      ),
+      ],
     );
   }
 }
