+++ lib/views/screens/welcome/welcome_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:health_care/common/app_colors.dart';
-import '../auth/Login/newLogin.dart';
+import '../auth/Login/login_screen.dart';
 
 class WelcomeScreen extends StatefulWidget {
   const WelcomeScreen({super.key});
